pyproject.toml
[project]
name = "full-pun-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "datasets>=3.2.0",
    "editdistance>=0.8.1",
    "flask-cors>=5.0.0",
    "flask>=3.1.0",
    "full-pun-generation",
    "gensim>=4.3.3",
    "gradio>=4.44.1",
    "huggingface-hub>=0.27.0",
    "keybert>=0.8.5",
    "langchain-community>=0.3.16",
    "langchain-ollama>=0.2.3",
    "langchain>=0.3.16",
    "nltk>=3.9.1",
    "phonemizer>=3.3.0",
    "polars>=1.19.0",
    "protobuf>=5.29.3",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=3.3.1",
    "sentencepiece>=0.2.0",
    "spacy==3.8.3",
    "tiktoken>=0.8.0",
    "torch>=2.5.1",
    "transformers[torch]>=4.47.0",
    "wandb>=0.19.6",
]

[project.scripts]
gradio = "full_pun_generation.interface:main"
pronunciation = "full_pun_generation.pronunciation:main"
test-wordnet = "full_pun_generation.wordnet:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "autopep8>=2.3.2",
    "pip>=24.3.1",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
full-pun-generation = { workspace = true }
